Code quality pass
